requirements.txt
Flask==2.3.4
Flask-Login==0.6.3
Flask-WTF==1.1.1
Werkzeug==2.3.6
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.4
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
python-dotenv==1.0.0
